Centralizes validation logic in service layer; adds email uniqueness.
