+++ src/main/java/ua/knd11/model/User.java
@@ -2,6 +2,9 @@ package ua.knd11.model;
 
 import java.util.Objects;
 
+import static ua.knd11.service.impl.UserServiceImpl.credentialsValidation;
+import static ua.knd11.service.impl.UserServiceImpl.normalizer;
+
 public abstract class User {
     private static int nextId = 1;
     private int id = 0;
@@ -13,34 +16,8 @@ public abstract class User {
     public User(String name, String surname, String email, String password) {
         this.name = normalizer(name, "Name");
         this.surname = normalizer(surname, "Surname");
-        this.email = credentialsValidation(email, "Email");
-        this.password = credentialsValidation(password, "Password");
-    }
-
-    public static String normalizer(String string, String type) {
-        Objects.requireNonNull(string, type + " must not be null");
-        if (string.isBlank()) throw new IllegalArgumentException(type + " must not be blank");
-
-        String regex = "^(?=.{1,100}$)[A-Za-zА-Яа-яЁёЄєҐґЇїІі0-9_-]+$";
-        if (!string.matches(regex)) throw new IllegalArgumentException("invalid " + type);
-
-        string = string.trim();
-        return string;
-    }
-
-    public String credentialsValidation(String credential, String type) {
-        if (type.equals("Email")) {
-            String regex = "^(?=.{1,254}$)(?=.{1,64}@)[A-Za-z0-9._%+-]+@[A-Za-z0-9.-]+\\.[A-Za-z]{2,}$";
-            if (!credential.matches(regex))
-                throw new IllegalArgumentException("Invalid email address");
-        }
-
-        if (type.equals("Password")) {
-            String regex = "^[-!@#$%^&*.A-Za-z\\d]{8,}$";
-            if (!credential.matches(regex))
-                throw new IllegalArgumentException("Invalid password");
-        }
-        return credential;
+        this.email = Objects.requireNonNull(credentialsValidation(email, "Email"));
+        this.password = Objects.requireNonNull(credentialsValidation(password, "Password"));
     }
 
     public void assignId() {
