+++ src/main/java/ua/knd11/security/impl/AuthServiceImpl.java
@@ -11,12 +11,6 @@ public class AuthServiceImpl implements AuthService {
 
     public void registration(User user) {
         if (user == null) throw new IllegalArgumentException("User must not be null");
-
-        for (User registered : registeredUsers) {
-            if (registered.getEmail().equalsIgnoreCase(user.getEmail())) {
-                throw new IllegalArgumentException("User with this email already exists");
-            }
-        }
         user.setPassword(String.valueOf(user.getPassword().hashCode()));
         registeredUsers.add(user);
     }
